requirements.txt
aiopg==1.3.5
asgiref==3.5.2
async-timeout==4.0.3
autopep8==2.0.4
certifi==2023.7.22
channels==4.0.0
channels-postgres==0.0.5
channels-redis==4.1.0
charset-normalizer==3.3.0
Django==4.1.12
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
msgpack==1.0.7
# psycopg2==2.9.8
psycopg2-binary==2.9.9
pycodestyle==2.11.0
PyJWT==2.8.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.6